Update project code - update some classes - refactoring

+++ src/main/java/it/unicam/cs/ids/services/ContenutoService.java
@@ -1,11 +1,13 @@
 package it.unicam.cs.ids.services;
 
 import it.unicam.cs.ids.dto.ContenutoDto;
+import it.unicam.cs.ids.enumeration.TipoCategoriaContenuto;
 import it.unicam.cs.ids.enumeration.TipoContenuto;
 import it.unicam.cs.ids.model.POI.POI;
 import it.unicam.cs.ids.model.POI.contenuto.Contenuto;
 import it.unicam.cs.ids.model.Users;
 import it.unicam.cs.ids.repository.ContenutoRepository;
+import it.unicam.cs.ids.repository.POIRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -18,23 +20,26 @@ public class ContenutoService {
     private ContenutoRepository contenutoRepository;
 
     @Autowired
-    private POIService poiService;
+    private POIRepository poiRepository;
 
     @Autowired
     private UserService userService;
 
 
     public Contenuto create(ContenutoDto contenutoDto) {
-        POI poi = poiService.read(contenutoDto.getPoiId());
+        POI poi = poiRepository.findById(contenutoDto.getPoiId())
+                .orElseThrow(() -> new RuntimeException("POI non trovato."));
         Users autore = userService.read(contenutoDto.getAutoreId());
 
         Contenuto contenuto;
+        TipoContenuto tipoContenuto = TipoContenuto.valueOf(contenutoDto.getTipo());
 
-        switch (contenutoDto.getTipo()) {
+        switch (tipoContenuto) {
             case TipoContenuto.TEMPORIZZATO:
                 contenuto = new Contenuto.Builder()
                         .setPOI(poi)
-                        .setTipo(contenutoDto.getTipo())
+                        .setTipo(tipoContenuto)
+                        .setCategoria(TipoCategoriaContenuto.valueOf(contenutoDto.getCategoria().toUpperCase()))
                         .setAutore(autore)
                         .setTitolo(contenutoDto.getTitolo())
                         .setDescrizione(contenutoDto.getDescrizione())
@@ -42,21 +47,24 @@ public class ContenutoService {
                         .setDataInizio(contenutoDto.getDataInizio())
                         .setDataFine(contenutoDto.getDataFine())
                         .setNote(contenutoDto.getNote())
+                        .setValidato(userService.getAuthenticatedUser().getAutorizzato()) // basandosi sui dettagli dell'utente autenticato preso dalla sessione
                         .build();
                 contenutoRepository.save(contenuto);
             case TipoContenuto.STATICO:
                 contenuto = new Contenuto.Builder()
                         .setPOI(poi)
-                        .setTipo(contenutoDto.getTipo())
+                        .setTipo(tipoContenuto)
+                        .setCategoria(TipoCategoriaContenuto.valueOf(contenutoDto.getCategoria().toUpperCase()))
                         .setAutore(autore)
                         .setTitolo(contenutoDto.getTitolo())
                         .setDescrizione(contenutoDto.getDescrizione())
                         .setUrl(contenutoDto.getUrl())
                         .setNote(contenutoDto.getNote())
+                        .setValidato(userService.getAuthenticatedUser().getAutorizzato()) // basandosi sui dettagli dell'utente autenticato preso dalla sessione
                         .build();
                 contenutoRepository.save(contenuto);
             default:
-                throw new IllegalArgumentException("Tipo contenuto non valido.");
+                throw new RuntimeException("tipo contenuto non valido.");
         }
     }
 
@@ -69,6 +77,8 @@ public class ContenutoService {
         Contenuto contenuto = contenutoRepository.findById(id)
                 .orElseThrow(() -> new RuntimeException("contenuto non trovato."));
 
+        contenuto.setTipo(TipoContenuto.valueOf(contenutoDto.getTipo().toUpperCase()));
+        contenuto.setCategoria(TipoCategoriaContenuto.valueOf(contenutoDto.getCategoria().toUpperCase()));
         contenuto.setTitolo(contenutoDto.getTitolo());
         contenuto.setDescrizione(contenutoDto.getDescrizione());
         contenuto.setUrl(contenutoDto.getUrl());
@@ -80,13 +90,23 @@ public class ContenutoService {
         return contenuto;
     }
 
-
     public void delete(Long id) {
-        contenutoRepository.deleteById(id);
+        if(contenutoRepository.existsById(id)) {
+            contenutoRepository.deleteById(id);
+        } else {
+            throw new RuntimeException("contenuto non trovato.");
+        }
     }
 
     public List<Contenuto> getAllContenutiByPOI(Long poiId) {
-        POI poi = poiService.read(poiId);
-        return contenutoRepository.findContenutoByPoi(poi);
+        return contenutoRepository.findContenutoByPoi(poiRepository.findById(poiId)
+                .orElseThrow(() -> new RuntimeException("POI non trovato.")));
+    }
+
+    public Contenuto validate(Long contenutoId) {
+        return contenutoRepository.findById(contenutoId).map(c -> {
+            c.setValidato(true);
+            return contenutoRepository.save(c);
+        }).orElseThrow(() -> new RuntimeException("contenuto non trovato."));
     }
 }
